requirements.txt
openai==1.12.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
pygit2==1.14.1
